requirements.txt
aiogram==3.10.0
yt-dlp==2025.9.5
python-dotenv==1.0.1
